+++ cloudsmith_cli/core/api/repos.py
@@ -4,7 +4,8 @@ from __future__ import absolute_import, print_function, unicode_literals
 import cloudsmith_api
 
 from . import user
-from .. import pagination, utils
+from .. import utils
+from ..pagination import PageInfo
 from .exceptions import catch_raise_api_exception
 from .init import get_api_client
 
@@ -34,5 +35,5 @@ def list_repos(owner=None, **kwargs):
             owner=owner, **api_kwargs
         )
 
-    page_info = pagination.get_page_info_from_headers(headers)
+    page_info = PageInfo.from_headers(headers)
     return [x.to_dict() for x in res], page_info